tests/test_app_update.py
import pytest
import os
import sys
from src.app_update import AppUpdate

EXPECTED_OLD_VERSION_STRING = "0307"

@pytest.fixture
def app_update():
    return AppUpdate()

@pytest.fixture
def invalid_search_location():
    return "http://invalid_location"

@pytest.fixture
def valid_search_location_old():
    return "https://api.github.com/repos/bstaple1/MTGA_Draft_17Lands/releases/96565628"

@pytest.fixture
def valid_search_location_latest():
    return "https://api.github.com/repos/bstaple1/MTGA_Draft_17Lands/releases/latest"

@pytest.fixture
def invalid_input_url():
    return "http://invalid_url"

@pytest.fixture
def valid_input_url_zip():
    return "https://github.com/bstaple1/MTGA_Draft_17Lands/releases/download/MTGA_Draft_Tool_V0307/MTGA_Draft_Tool_V0307_Setup.zip"
    
@pytest.fixture
def valid_input_url_exe():
    return "https://github.com/bstaple1/MTGA_Draft_17Lands/releases/download/MTGA_Draft_Tool_V0304/MTGA_Draft_Tool_V0304_Setup.exe"

@pytest.fixture
def output_filename():
    return "test_file.exe"
    
def test_retrieve_file_version_latest_success(app_update, valid_search_location_latest):
    version, file_location = app_update.retrieve_file_version(valid_search_location_latest)
    assert isinstance(version, str) and isinstance(float(version), float)
    assert isinstance(file_location, str)
    assert len(file_location) != 0
    
def test_retrieve_file_version_old_success(app_update, valid_search_location_old):
    version, file_location = app_update.retrieve_file_version(valid_search_location_old)
    assert version == EXPECTED_OLD_VERSION_STRING
    
def test_retrieve_file_version_failure(app_update, invalid_search_location):
    version, file_location = app_update.retrieve_file_version(invalid_search_location)
    assert version == ""
    assert file_location == ""

@pytest.mark.skipif(sys.platform != "win32", reason="Windows only")
def test_download_file_zip_success(app_update, valid_input_url_zip, output_filename):
    output_location = app_update.download_file(valid_input_url_zip, output_filename)
    assert isinstance(output_location, str) and os.path.exists(output_location)
    
@pytest.mark.skipif(sys.platform != "win32", reason="Windows only")
def test_download_file_exe_success(app_update, valid_input_url_exe, output_filename):
    output_location = app_update.download_file(valid_input_url_exe, output_filename)
    assert isinstance(output_location, str) and os.path.exists(output_location)

@pytest.mark.skipif(sys.platform != "win32", reason="Windows only")
def test_download_file_failure(app_update, invalid_input_url, output_filename):
    output_location = app_update.download_file(invalid_input_url, output_filename)
    assert output_location == ""
